requirement.txt
asgiref==3.7.2
click==8.1.4
cmake==3.26.4
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
numpy==1.25.1
Pillow==10.0.0
PyJWT==2.7.0
pytz==2023.3
simplejson==3.19.1
sqlparse==0.4.4
